reorganized for better readability and maintainability
