+++ model.json
@@ -1,7 +1,7 @@
 {
     "model_with_ferplus" : 
         {
-        "modelPath" : "clfs/model_with_ferplus.h5",
+        "modelPath" : "trained_models/model_with_ferplus.h5",
         "inputSize" : [48, 48, 1],
         "labels" : ["neutral", "happiness", "surprise", "sadness", "anger", "disgust", "fear", "contempt", "unknown", "NF"]
         }
